testing out hierarchical Fsm
Added test for a simple hierarchical Fsm with 2 subFsm modules. This version only separates on the addTransition level for protocol construction and imports Transitions from children to parent. Actual handling of Transitions are done in the same way as for BisqEasy.

Goal was to explore possible solutions, so priority has not been given to security and privacy concerns at this stage.

For an actual implementation of this approach, the TradeStates would still need to be defined in a similar way to BisqEasyTradeState to ensure ordinal sequence for events.

Question still remains if separation of concerns should be extended past the protocol level.